ISPN-8058 Move org.infinispan.tasks.Task to the tasks-api JAR
- Also move some methods from ServerTask to Task
- Make ServerTask implement Task
